corrección de errores

+++ app/database/migrations/2015_05_17_013416_create_establishments_table.php
@@ -16,8 +16,8 @@ class CreateEstablishmentsTable extends Migration {
 		{
 			$table->increments('id');
 			$table->string('name');
-			$table->double('lat',2,12);
-			$table->double('lng',2,12);
+			$table->float('lat');
+			$table->float('lng');
 		});
 	}
